Holy queue ownership transfers, Batman!
Okay so I messed up: even in exclusive queue ownership mode AND with otherwise valid memory barriers it is REQUIRED to do ownership transfers. Clients must now specify the queue they intend to use their resources on, when giving them initial data. This is because the first use acquires the object for that queue (transfer), so we need to make sure to change it back to what the client wants for validation reasons.