Fix ResourceLoader resource duplication and Stop() handling
Also improve usage of mutexes. We now make less assumptions about the passed name - not forcing it to be an absolute file path. Additionally, I change the stop() function to correctly close pending work and publish it - without erasing this work. We save release for the final dereferencing of the object, and I made sure to help make it easier to fully release resources.