Need to finish gluing this together buuut... extension wrangler!
Currently working on the python script side of things. We'll run this as part of the build pipeline via CMake, so that before the renderer context is generated we generate a header of a ton of extension lookup data. Extension wrangler will use this to make for easy and reliable queries of extensions, and can provide the ability to enable dependencies automatically (and also tracks what extensions were made core parts of the API)