Iterate more on extension python script + some missed changes
Change how feature struct generation works to work upfront with the extensions themselves, also taking the chance to filter out provisional extensions and platform-specific extensions. Will need to return to tweak this to actually map to the current execution platform of the script, and would be nice to see if we can't change the provisional extension behavior based on this as well.