Add "endFrameFences" to VulkanScene
At the start of each frame, we wait on this fence - created initially in the signaled state for the first wait, and then used to make sure all previous work tied to a frame's contextual data has completed before proceeding. This lets the other N frames in the round-robin buffer run freely until we get around to this point.