Fix try_pop() function in mwsrQeue
The startRead() function used by the exit reactor makes the assumption - and a reasonable one at that given how this implementation works - that we're going to lock if we can't read. We can fix this by adding a tryStartRead() function, which explicitly avoids locking but otherwise behaves the same. We'd run into an assert about the reader being locked, because try_pop() keeps retrying to read without waiting to be woken/unlocked. Neat!