Replace mutex+CV usage in MWSR Queue with atomic_flag and C++20 wait
We can now do an atomic wait and notify operation like a mutex+condition_variable pair with atomics, if the feature is supported. LockedSingleThread always just has a value of 0 or 1 for it's counter, so we can switch to a simpler flag and have the thread wait on that - writers can wake it by changing the value and signaling. ifdef'd functionality just in case, too - will use a spin otherwise