Fix sneaky lifetime bug in ResourceContext
In the case of pointer members in the resource create info structs, they end up being destroyed before we create the resource in most cases because we exit the scope they were declared in while passing the work off to the other thread. Instead, we need to immediately translate the Vk info structs from the vulkan type to our internal type during the process of pushing a message to the queue. This will perform a deep copy of all pointer members, and then later we convert back to the vk info structures as needed.