Implementing missing swapchain functions, iterating on HDR impl
Added a bunch of unimplemented functions in the interface class, and changed how we handle surface format selection and HDR setup. Now we try and select a best format from a collection of vaguely scored formats, and when provided with a desired format we try our best to match that with what is currently supported.